Projects/edupath/edupath/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
razorpay==1.4.2
neo4j==5.15.0
openai==1.10.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
pyjwt==2.8.0
bleach==6.1.0
python-jose==3.3.0
passlib==1.7.4


pypdf2==3.0.1
python-docx==1.1.0
pytube==15.0.0
youtube-transcript-api==0.6.1
sentence-transformers==2.2.2
chromadb==0.4.22
langchain==0.1.0
tiktoken==0.5.2

# ✅ ADD: For Knowledge Graph RAG
spacy==3.7.2
google-search-results==2.4.2  # SerpAPI
flask-mail==0.10.1